requirements.txt
absl-py==1.4.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==22.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.26.104
botocore==1.29.104
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
filelock==3.10.4
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
flatbuffers==23.3.3
fonttools==4.39.2
idna==3.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.11.0
kiwisolver==1.4.4
labelImg==1.8.6
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe-silicon==0.9.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.1
Pillow==9.4.0
platformdirs==3.1.1
protobuf==3.19.4
pyparsing==3.0.9
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pywatchman==1.4.1
redis==5.0.0
requests==2.31.0
s3transfer==0.6.0
scipy==1.10.1
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
sklearn==0.0.post4
soupsieve==2.4.1
spotipy==2.23.0
SQLAlchemy==2.0.22
sqlparse==0.4.3
sympy==1.11.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.3.7
